Adicionar as duas regras de negocio, um aluno só pode avaliar professores da sua turma e não pode avaliar professores em menus de uma semana

+++ src/main/java/dm/java10x/AvaliacaoDeProfessores/repository/TurmaRepository.java
@@ -8,5 +8,5 @@ import java.util.List;
 
 public interface TurmaRepository extends JpaRepository<TurmaModel, Long> {
 
-    List<Integer> findId_professorByTurma(Turma turma);
+    List<Long> findId_professorByTurma(Turma turma);
 }
